+++ src/Plugins/SubjectExecutor.php
@@ -26,6 +26,8 @@ use TechDivision\Import\Observers\ObserverVisitorInterface;
 use TechDivision\Import\Subjects\SubjectFactoryInterface;
 use TechDivision\Import\Subjects\ExportableSubjectInterface;
 use TechDivision\Import\Configuration\SubjectConfigurationInterface;
+use League\Event\EmitterInterface;
+use TechDivision\Import\Utils\EventNames;
 
 /**
  * The subject executor instance.
@@ -61,22 +63,32 @@ class SubjectExecutor implements SubjectExecutorInterface
     protected $subjectFactory;
 
     /**
+     * The event emitter instance.
+     *
+     * @var \League\Event\EmitterInterface
+     */
+    protected $emitter;
+
+    /**
      * Initializes the plugin with the application instance.
      *
      * @param \TechDivision\Import\Callbacks\CallbackVisitorInterface $callbackVisitor The callback visitor instance
      * @param \TechDivision\Import\Observers\ObserverVisitorInterface $observerVisitor The observer visitor instance
      * @param \TechDivision\Import\Subjects\SubjectFactoryInterface   $subjectFactory  The subject factory instance
+     * @param \League\Event\EmitterInterface                          $emitter         The event emitter instance
      */
     public function __construct(
         CallbackVisitorInterface $callbackVisitor,
         ObserverVisitorInterface $observerVisitor,
-        SubjectFactoryInterface $subjectFactory
+        SubjectFactoryInterface $subjectFactory,
+        EmitterInterface $emitter
     ) {
 
         // initialize the callback/observer visitors
         $this->callbackVisitor = $callbackVisitor;
         $this->observerVisitor = $observerVisitor;
         $this->subjectFactory = $subjectFactory;
+        $this->emitter = $emitter;
     }
 
     /**
@@ -96,6 +108,14 @@ class SubjectExecutor implements SubjectExecutorInterface
         $subjectInstance = $this->subjectFactory->createSubject($subject);
 
         try {
+            // load the subject + plug-in ID to prepare the event names
+            $subjectName = $subject->getName();
+            $pluginName = $subject->getPluginConfiguration()->getName();
+
+            // invoke the event that has to be fired before the subject's import method will be invoked
+            $this->emitter->emit(EventNames::SUBJECT_IMPORT_START, $subjectInstance);
+            $this->emitter->emit(sprintf('%s.%s.%s', $pluginName, $subjectName, EventNames::SUBJECT_IMPORT_START), $subjectInstance);
+
             // setup the subject instance
             $subjectInstance->setUp($serial);
 
@@ -108,20 +128,46 @@ class SubjectExecutor implements SubjectExecutorInterface
 
             // query whether or not, we've to export artefacts
             if ($subjectInstance instanceof ExportableSubjectInterface) {
-                $subjectInstance->export(
-                    $matches[BunchKeys::FILENAME],
-                    $matches[BunchKeys::COUNTER]
-                );
+                try {
+                    // invoke the event that has to be fired before the subject's export method will be invoked
+                    $this->emitter->emit(EventNames::SUBJECT_EXPORT_START, $subjectInstance);
+                    $this->emitter->emit(sprintf('%s.%s.%s', $pluginName, $subjectName, EventNames::SUBJECT_EXPORT_START), $subjectInstance);
+
+                    // export the artefacts if available
+                    $subjectInstance->export(
+                        $matches[BunchKeys::FILENAME],
+                        $matches[BunchKeys::COUNTER]
+                    );
+
+                    // invoke the event that has to be fired after the subject's export method has been invoked
+                    $this->emitter->emit(EventNames::SUBJECT_EXPORT_SUCCESS, $subjectInstance);
+                    $this->emitter->emit(sprintf('%s.%s.%s', $pluginName, $subjectName, EventNames::SUBJECT_EXPORT_SUCCESS), $subjectInstance);
+                } catch (\Exception $e) {
+                    // invoke the event that has to be fired when the subject's export method throws an exception
+                    $this->emitter->emit(EventNames::SUBJECT_EXPORT_FAILURE, $subjectInstance);
+                    $this->emitter->emit(sprintf('%s.%s.%s', $pluginName, $subjectName, EventNames::SUBJECT_EXPORT_FAILURE), $subjectInstance);
+
+                    // re-throw the exception
+                    throw $e;
+                }
             }
 
             // tear down the subject instance
             $subjectInstance->tearDown($serial);
+
+            // invoke the event that has to be fired after the subject's import method has been invoked
+            $this->emitter->emit(EventNames::SUBJECT_IMPORT_SUCCESS, $subjectInstance);
+            $this->emitter->emit(sprintf('%s.%s.%s', $pluginName, $subjectName, EventNames::SUBJECT_IMPORT_SUCCESS), $subjectInstance);
         } catch (\Exception $e) {
             // query whether or not, we've to export artefacts
             if ($subjectInstance instanceof ExportableSubjectInterface) {
                 // tear down the subject instance
                 $subjectInstance->tearDown($serial);
             }
+
+            // invoke the event that has to be fired when the subject's import method throws an exception
+            $this->emitter->emit(EventNames::SUBJECT_IMPORT_FAILURE, $subjectInstance);
+            $this->emitter->emit(sprintf('%s.%s.%s', $pluginName, $subjectName, EventNames::SUBJECT_IMPORT_FAILURE), $subjectInstance);
 
             // re-throw the exception
             throw $e;